Add a remote-only Lix entrypoint
