+++ packages/js-sdk/test-fixtures/vite-remote/index.html
@@ -1,0 +1,11 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<title>Lix SDK remote bundle smoke</title>
+	</head>
+	<body>
+		<script type="module" src="/src/main.js"></script>
+	</body>
+</html>
