+++ packages/js-sdk/test-fixtures/vite-remote/src/main.js
@@ -1,0 +1,3 @@
+import { openLix } from "@lix-js/sdk/remote";
+
+globalThis.__openRemoteLix = openLix;
